refactor: ヘッダーのサイドバーのコンポーネント構造の変更
