--- src/layouts/SideDrawer.vue
@@ -1,9 +1,0 @@
-<script lang="ts" setup>
-import SideNavigation from '@/layouts/SideNavigation.vue'
-</script>
-
-<template>
-  <div class="fixed left-0 top-12 h-full w-1/2 bg-background-default-default pt-4">
-    <SideNavigation />
-  </div>
-</template>
